+++ UnitTests/BraintreeThreeDSecureTests/BTThreeDSecure_Tests.swift
@@ -14,37 +14,40 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         threeDSecureRequest = BTThreeDSecureRequest()
         threeDSecureRequest.amount = 10.0
         threeDSecureRequest.nonce = "fake-card-nonce"
-        threeDSecureRequest.versionRequested = .version1
+        threeDSecureRequest.versionRequested = .version2
         mockAPIClient = MockAPIClient(authorization: tempClientToken)!
         mockThreeDSecureRequestDelegate = MockThreeDSecureRequestDelegate()
+        threeDSecureRequest.threeDSecureRequestDelegate = mockThreeDSecureRequestDelegate
     }
 
     // MARK: - ThreeDSecure Authentication Tests
+    
+    func testStartPayment_whenV1Attempted_returnsError() {
+        threeDSecureRequest.versionRequested = .version1
+        threeDSecureRequest.threeDSecureRequestDelegate = mockThreeDSecureRequestDelegate
 
-    func testStartPayment_displaysSafariViewControllerWhenAvailable_andRequiresAuthentication() {
-        BTAppContextSwitcher.setReturnURLScheme("com.braintreepayments.Demo.payments")
-        
-        let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
-        
-        viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
-        
+        let expectation = self.expectation(description: "willCallCompletion")
+
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
+            "threeDSecure": [],
             "assetsUrl": "http://assets.example.com"
         ])
         let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
-
-        mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
 
         driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
-            
+            XCTAssertNotNil(error)
+            XCTAssertNil(result)
+            guard let error = error as NSError? else {return}
+            XCTAssertEqual(error.domain, BTThreeDSecureFlowErrorDomain)
+            XCTAssertEqual(error.code, BTThreeDSecureFlowErrorType.configuration.rawValue)
+            XCTAssertEqual(error.localizedDescription, "3D Secure v1 is deprecated and no longer supported. See https://developer.paypal.com/braintree/docs/guides/3d-secure/client-side for more information.")
+            expectation.fulfill()
         }
-        
+
         waitForExpectations(timeout: 4, handler: nil)
     }
 
     func testStartPayment_v2_returnsErrorWhenCardinalAuthenticationJWT_isMissing() {
-        threeDSecureRequest.versionRequested = .version2
         threeDSecureRequest.threeDSecureRequestDelegate = mockThreeDSecureRequestDelegate
 
         let expectation = self.expectation(description: "willCallCompletion")
@@ -155,6 +158,13 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecureInfo":     [
                 "liabilityShiftPossible": false,
                 "liabilityShifted": false,
+            ],
+            "lookup": [
+                "acsUrl": nil,
+                "pareq": "",
+                "md": "",
+                "termUrl": "http://example.com",
+                "threeDSecureVersion": "2.0"
             ]
             ] as [String : Any]
         mockAPIClient.cannedResponseBody = BTJSON(value: responseBody)
@@ -200,6 +210,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
     func testStartPayment_v2_when_threeDSecureRequestDelegate_notSet_returnsError() {
         let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
         threeDSecureRequest.versionRequested = .version2
+        threeDSecureRequest.threeDSecureRequestDelegate = nil
 
         let expectation = self.expectation(description: "willCallCompletion")
 
@@ -225,27 +236,21 @@ class BTThreeDSecure_UnitTests: XCTestCase {
 
     func testStartPayment_successfulResult_callsCompletionBlock() {
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
+            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com",
         ])
         
-        let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
-        viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
-        
         let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
 
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
         
-        var paymentFinishedExpectation: XCTestExpectation? = nil
+        let paymentFinishedExpectation = expectation(description: "Start payment expectation")
         driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNil(error)
             XCTAssertNotNil(result)
-            paymentFinishedExpectation!.fulfill()
+            paymentFinishedExpectation.fulfill()
         }
         
-        waitForExpectations(timeout: 2, handler: nil)
-        
-        paymentFinishedExpectation = self.expectation(description: "Start payment expectation")
         BTPaymentFlowDriver.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22paymentMethod%22:%7B%22type%22:%22CreditCard%22,%22nonce%22:%220d3e1cc8-50a4-0437-720b-c03c722f0d0a%22,%22description%22:%22ending+in+02%22,%22consumed%22:false,%22threeDSecureInfo%22:%7B%22liabilityShifted%22:true,%22liabilityShiftPossible%22:true,%22status%22:%22authenticate_successful%22,%22enrolled%22:%22Y%22%7D,%22details%22:%7B%22lastTwo%22:%2202%22,%22lastFour%22:%220002%22,%22cardType%22:%22Visa%22%7D,%22bin_data%22:%7B%22prepaid%22:%22Unknown%22,%22healthcare%22:%22Unknown%22,%22debit%22:%22Unknown%22,%22durbin_regulated%22:%22Unknown%22,%22commercial%22:%22Unknown%22,%22payroll%22:%22Unknown%22,%22issuing_bank%22:%22Unknown%22,%22country_of_issuance%22:%22Unknown%22,%22product_id%22:%22Unknown%22%7D%7D,%22threeDSecureInfo%22:%7B%22liabilityShifted%22:true,%22liabilityShiftPossible%22:true%7D,%22success%22:true%7D")!)
         
         waitForExpectations(timeout: 2, handler: nil)
@@ -253,30 +258,24 @@ class BTThreeDSecure_UnitTests: XCTestCase {
     
     func testStartPayment_returnsFailedAuthenticationError_whenErrorReturnedInURL() {
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
+            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com",
             ])
         
-        let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
-        viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
-        
         let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
-        
+
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
         
-        var paymentFinishedExpectation: XCTestExpectation? = nil
+        let paymentFinishedExpectation = expectation(description: "Start payment expectation")
         driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
             guard let error = error as NSError? else {return}
             XCTAssertEqual(error.domain, BTThreeDSecureFlowErrorDomain)
             XCTAssertEqual(error.code, BTThreeDSecureFlowErrorType.failedAuthentication.rawValue)
-            paymentFinishedExpectation!.fulfill()
+            paymentFinishedExpectation.fulfill()
         }
-        
-        waitForExpectations(timeout: 2, handler: nil)
-        
-        paymentFinishedExpectation = self.expectation(description: "Start payment expectation")
+
         BTPaymentFlowDriver.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22threeDSecureInfo%22:%7B%22liabilityShifted%22:false,%22liabilityShiftPossible%22:true%7D,%22error%22:%7B%22message%22:%22Failed+to+authenticate,+please+try+a+different+form+of+payment.%22%7D,%22success%22:false%7D")!)
         
         waitForExpectations(timeout: 2, handler: nil)
@@ -284,18 +283,14 @@ class BTThreeDSecure_UnitTests: XCTestCase {
 
     func testStartPayment_missingAuthResponse_callsCompletionBlock_withError_sendsAnalyticsEvent() {
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
+            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com",
             ])
 
-        let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
-        viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
-
         let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
-
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
 
-        var paymentFinishedExpectation: XCTestExpectation? = nil
+        let paymentFinishedExpectation = expectation(description: "Start payment expectation")
         driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
@@ -303,12 +298,9 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             XCTAssertEqual(error.domain, BTThreeDSecureFlowErrorDomain)
             XCTAssertEqual(error.code, BTThreeDSecureFlowErrorType.failedAuthentication.rawValue)
             XCTAssertEqual(error.localizedDescription, "Auth Response missing from URL.")
-            paymentFinishedExpectation!.fulfill()
+            paymentFinishedExpectation.fulfill()
         }
 
-        waitForExpectations(timeout: 2, handler: nil)
-
-        paymentFinishedExpectation = self.expectation(description: "Start payment expectation")
         BTPaymentFlowDriver.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?no-auth=bad-response")!)
 
         waitForExpectations(timeout: 2, handler: nil)
@@ -318,18 +310,15 @@ class BTThreeDSecure_UnitTests: XCTestCase {
 
     func testStartPayment_invalidAuthResponse_callsCompletionBlock_withError_sendsAnalyticsEvent() {
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
+            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com",
             ])
 
-        let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
-        viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
-
         let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
 
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
 
-        var paymentFinishedExpectation: XCTestExpectation? = nil
+        let paymentFinishedExpectation = expectation(description: "Start payment expectation")
         driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
@@ -337,12 +326,9 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             XCTAssertEqual(error.domain, BTThreeDSecureFlowErrorDomain)
             XCTAssertEqual(error.code, BTThreeDSecureFlowErrorType.failedAuthentication.rawValue)
             XCTAssertEqual(error.localizedDescription, "Auth Response JSON parsing error.")
-            paymentFinishedExpectation!.fulfill()
+            paymentFinishedExpectation.fulfill()
         }
 
-        waitForExpectations(timeout: 2, handler: nil)
-
-        paymentFinishedExpectation = self.expectation(description: "Start payment expectation")
         BTPaymentFlowDriver.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22paymentMethod%22:%7B%22type%22:%22CreditCard%22,%22nonce%22:%220d3e1cc8-50a4-0437-720b-c03c722f0d0a%22,BAD-JSON")!)
 
         waitForExpectations(timeout: 2, handler: nil)
@@ -352,18 +338,15 @@ class BTThreeDSecure_UnitTests: XCTestCase {
 
     func testStartPayment_unexpectedAuthResponse_callsCompletionBlock_withError_sendsAnalyticsEvent() {
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
+            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com",
             ])
 
-        let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
-        viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
-
         let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
 
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
 
-        var paymentFinishedExpectation: XCTestExpectation? = nil
+        let paymentFinishedExpectation = expectation(description: "Start payment expectation")
         driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
@@ -371,12 +354,9 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             XCTAssertEqual(error.domain, BTThreeDSecureFlowErrorDomain)
             XCTAssertEqual(error.code, BTThreeDSecureFlowErrorType.failedAuthentication.rawValue)
             XCTAssertEqual(error.localizedDescription, "Auth Response JSON parsing error.")
-            paymentFinishedExpectation!.fulfill()
+            paymentFinishedExpectation.fulfill()
         }
 
-        waitForExpectations(timeout: 2, handler: nil)
-
-        paymentFinishedExpectation = self.expectation(description: "Start payment expectation")
         BTPaymentFlowDriver.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%22STRING%22")!)
 
         waitForExpectations(timeout: 2, handler: nil)
@@ -411,7 +391,9 @@ class BTThreeDSecure_UnitTests: XCTestCase {
                 "acsUrl": "http://example.com",
                 "pareq": "",
                 "md": "",
-                "termUrl": "http://example.com"
+                "termUrl": "http://example.com",
+                "threeDSecureVersion": "2.0",
+                "transactionId": "111222333"
             ]
         ]
     }
@@ -421,9 +403,8 @@ class BTThreeDSecure_UnitTests: XCTestCase {
     func testStartPayment_success_sendsAnalyticsEvents() {
         let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
 
-        viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
-
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
+            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com",
             ])
         let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
@@ -460,8 +441,9 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             ] as [String : Any]
         mockAPIClient.cannedResponseBody = BTJSON(value: responseBody)
 
+        let expectation = expectation(description: "Sends all analytics events")
         driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
-
+            expectation.fulfill()
         }
 
         waitForExpectations(timeout: 4, handler: nil)
@@ -470,11 +452,11 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains("ios.three-d-secure.verification-flow.started"))
         XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains("ios.three-d-secure.verification-flow.3ds-version.1.0"))
         XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains("ios.three-d-secure.verification-flow.challenge-presented.true"))
-        XCTAssertTrue(mockAPIClient.postedAnalyticsEvents.contains("ios.three-d-secure.webswitch.initiate.succeeded"))
     }
 
     func testStartPayment_failure_sendsAnalyticsEvents() {
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
+            "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com",
             ])
         let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)